Filter out the "non-standard" tags coming as default_tags
Databricks adds a number of the default tags, like, Creator, ClusterName, ClusterId,
etc. that are returned as `default_tags` objects in the REST API Clusters Get call.  But
this block also returns the tags inherited from the Azure resource group, workspace, etc.
But this block is treated as Spark struct column, and this leads to addition of the
"non-standard" names that are becoming the fields of struct. But in case of mixing the
casing of the same name, the processing is failing when writing data into Delta table.

This fix doing two things:
* leaving only Databricks-specific tags in the `default_tags` column
* converting the original `default_tags` data coming from API into a separate column of
  the map type, so users will still able to query the data using the "non-standard" tags

this fixes #260